removed 'print' within app.py

+++ src/pipeline/predict_pipeline.py
@@ -44,8 +44,8 @@ class CustomData:
 		parental_level_of_education:str,
 		lunch:str,
 		test_preparation_course:str,
-		reading_score:int,
-		writing_score:int):
+		reading_score:float,
+		writing_score:float):
 
 		# assigining the values
 		self.gender = gender
